issue-15:  make vacuum go faster
First, we only test rows for deletion that might actually be dead (_xmax_is_committed:true or _xmin_is_committed:false).  This reduces the total number of rows we need to examine to what's likely the complete set, rather than needing to inspect each and every row in the index.  So instead of looking at 20M+ rows, we just look at the 50k that have been UPDATEd or DELETEd.

Secondly, when building the internal set of ItemPointers to actually delete, use a `List` instead of manually growing/copying a `char *`.

This also renames the "getAllItems" function to "getPossiblyExpiredItems" because that's what it does, and "get all items" isn't a thing we do anywhere.