fix issue #29:  Use the "max_terms" value for tally aggregate function
QueryRewriter.java was hardcoding zero as the .size() argument to the TermsBuilder, which means "get everything".  This change sets it to the value provided by the user in their call to zdb_tally().  A user-provided value of zero now means "get everything".

The .shardSize() argument is still hardcoded to zero, which means that ES will be as precise as it possibly can when pulling terms from each shard.

Thanks to @ryancutter for the report and discussion.

Also:
  - Aggregation execution time logging was stupidly logging before the aggregate actually finished.
  - a test for issue #29 to make sure results from zdb_tally() are indeed limited by the 'max_terms' argument.