working on issue #20:  first step was to change TestQueryRewriter.java so that's connected to a real Elasticsearch node (bootstrapped via ZomboDBTestCase.java -- not sure this is the best way to bootstrap an ES node, but wasn't able to figure out anything else).
This uncovered a bug where the _all field was never being put into the set of fields in IndexMetadataManager, and as such was never being detected as having an analyzer of type "phrase", which meant certain types of phrases weren't being properly converted into spans against the _all field.

There's a lot of churn in the unit tests 'expected' results, but that's because the framework is now connected.

All the SQL-level tests still pass without changes, so that's a good sign nothing fundamental was broken here.