move the process of doing expansions out of QueryRewriter and into a new class called ExpansionOptimizer and call it after we do all other optimizations.
After the nodes have been expanded, we then go through the tree and do our best to merge adjacent ASTArrays (the result of an expansion) when they're in an ASTAnd.

This drastically improves the performance of doing joins in typical use cases because we only need to search the intersection of the documents rather than all matching documents on all sides.