Fix a long standing bug where deleting docs (via vacuum) could, depending on timing when that ctid is reused by PG, cause primary and secondary replicas to become out of sync.
The fix for this is to use an explicit version (value of 2) when creating and deleting docs.

Additionally, set the OpType to CREATE when inserting new docs so we'll catch possible corruption issues earlier.

Also additionally, do some extra validation during the handleDeleteRequests() method to catch possible corruption issues earlier.