fix bug where we considered a VersionConflict an error during vacuum when deleting "xmax" docs.
This is actually a perfectly normal and expected response in cases
where we try to delete an "xmax" doc that's been replaced by a new doc
between when we thought we needed to delete it and when we actually do.

This really should have been part of 34377b946f052f679ccc3c5bbfc8cc6362da16df

Also, change logging message so that a tail postgresql.log | grep "zombodb vacuum"
will actually find this error.