directly copy in QueryRewriter.java and TestQueryRewriter.java from v3.2.0, and fix QueryRewriter.java so it compiles and is generally compatible with ES 2.4.
TestQueryRewriter lights up quite a few tests.  In general the
differences all look okay, but I'm not sure about the "regex"
queries.  With ES 2.4 there's a bunch of default flags set
that weren't set in ES 1.7.  I don't know if the meaning is the
same or not.  I need to investigate further before I fix everything
else that TestQueryRewriter complains about.