add new index and session-level configuration options for controlling how ZomboDB estimates the number of rows that'll be returned from a SELECT query.
Historically, ZomboDB would make a call to Elasticsearch to "_count" the
number of rows that'll likely be returned by the query being executed.
And that number would be used for query planning selectivity purposes.

Now, ZomboDB just defaults to 2500 rows.  This number was chosen as it
change any of the existing plans that ZDB's regression tests examine.

In general, when ZomboDB indexes are involved, it's always going to be
faster for Postgres to plan an Index Scan rather than a Sequential Scan,
and 2500 is a good number for that.

If more precise estimates are required, however, `USING zombodb` indexes
now have an option called "default_row_estimate".  Setting it to any positive
value will always return that value, setting it to `-1` will cause ZDB
to revert to its old behavior of directly asking Elasticsearch.

This behavior can also be forced on a case-by-case basis be setting the
new Postgres "GUC" called `zombodb.force_row_estimation` to `ON`.