implement a shard action called FastTerms
FastTerms is a shard action that given an index, a query, and a field
streams back the DocValues for that field from all docs that match
the query.  This has proven to be significantly faster than doing a
normal SearchAction and requesting the _zdb_id field.

FastTerms are used during normal SELECT statements when scores aren't
necessary and there's no #limit() clause.  This allows ZDB to stream
back the matching tuple ids very quickly.  3.2s to .220s to return 1M
rows.

It's now also used when ZomboDB needs to resolve cross-index joins via
index links.  This is also significantly faster than what ZDB used to do
and may actually be on-par with ES 1.7+SIREn -- I need to do more
benchmarking and likely more optimizations here.