cleanup how/when we sort results by _zdb_id
Getting the results sorted by _zdb_id (or postgres (block,offset)) is
really important for performance.  Now that we have FastTerms, we
can sort the results per shard in parallel and then merge-sort those
together when building the final bytes.

When we need to query ES using SearchAction we now ask it to sort
by _zdb_id for us, so we don't need to do it ourselves.

We're left only needing to quick-sort results in ZomboDBTIDResponseAction
only when a query has a #limit() clause.