Teach the CrossJoin stuff to use a java.util.BitSet when it makese sense
Generally, this is when the span of numbers we're trying to join with
fit within 32 bits.  This will always work for integers since they are
32bits, but for longs we see if (max_value - min_value) is within 32bits.
If so we scale them back to zero and then encode using a BitSet.

Otherwise, we'll stick to using sorted long[]s and binary searching the
arrays.

We could have a combination of the two from different shards, so all the
logic is encapsulated in a new class called NumberArrayLookup.java.

There might still be room for optimization here, but looking at ES'
/_nodes/hot_threads output with various intervals pretty much shows
we're back to spending time walking doc values inside Lucene, so that's
really positive.

(Outside of a different data structure and searching algorithm, it might
make sense to focus on reducing the amount of network bandwidth.  Right
now I'm just using Java serialization for the BitSets, and I'm not exactly
sure if that's more compact/efficient than, say, delta encoding a list
of longs and encoding as variable-width values across the network and then
decoding it back into a BitSet.)

On my test cluster, this is about 4x faster than before this commit
and maybe 2.5x faster than SIREn running on ES 1.7.