fixing issue #260
This changes how nested cross-join-queries are resolved such that they're
resolved, on each shard, from the bottom-up.  With nested cross-join-queries
only the top-level one can be optimized when block_routing_field is set...
all child cross-join-queries will always execute across all shards

1) Change the thread pool executor used in TransportFastTermsAction back
to GENERIC.  We need a threaded executor in order to cleanly make a
request back to the cluster.

2) Teach TransportFastTermsAction to only execute on the specific shard
it needs, if optimization is possible

3) During CrossJoinQueryBuilder's rewrite process we build a stack
of nested CrossJoinQueryBuilder objects, and get the FastTermsResponse for
each, starting at the bottom and working our way up to the top.  This ensures
that we don't have any reentrancy back into the cluster that could cause
the thread pool being used by TransportFastTermsAction to deadlock.

If the top-level one indicates that it can be optimized, then we delay
executing it until we transform it into a Query, so that we know which
shard we're executing on and can optimize for.

4) Add equals/hashcode methods to a few more classes

5) Change setup.sql to create way too many shards, and with block routing
so it's more likely to expose problems in the future