fixing issue #286
This necessitated rewriting how we find usages of zdb.score() in the
current query by walking the planstate and evaluating expressions.

It also required moving when we actually query ES in the index method
from amrescan() to amgettuple()/amgetbitmap() so that we're sure the
query has been fully planned.

In the case of sequential scans, any mention of zdb.score() against a
particular table will cause all ES queries for that table to require
scores, even if they're subqueries that don't need scores.