implement a test for issue #338.
In order to do this, I had to plumb through a new index-level setting
named "always_join_with_docvalues" (boolean), but I am not documenting
this feature.  It exists only for this particular test.

Additionally, issue #338 only shows itself on a multi-node cluster, so
I have the test hardcoded to reference my local multi-node cluster, and
have hacked it out of the tests list when the test suite runs on Travis.

I'm not happy about this, but here it is.