fix issue #29 for zdb-rust
This is implementing a new function `zdb.tally(...)` that's equivalent to the old `zdb_tally(...)` functions.  Required plumbing some support through our aggregate execution API client to execute a set of aggregates as `zdb.tally()` also wants to also always return the count of null values for the field.

Also updated a few column types for so_posts to be compatible with test expectations.