fix issue #272 for zdb-rust
This is a big changeset.  #272 exposed that ZDB-Rust wasn't properly handling nested queries.

We now pull up nested objects together into (potentially nested!) Expr::Nested nodes for each part of the nested path, providing proper conversion to DSL.

And all aggregate search functions now take the field name upon which they're aggregating, and we determine if that field is nested.  If it is, we build a "nested" aggregate that wraps a "filter" aggregate that wraps the actual aggregate.

The "filter" aggregate is only generated if the query itself contains any "nested" query clauses.  And the "filter" query is extracted from the generated ES QueryDSL json (so we can support arbitrary user queries and/or ZDB-query_parser queries).