fixes for issue #613
We now make use of ES' "index order" to order indexes by "zdb_ctid" (when the table doesn't have "type": "nested" mappings, which is an ES liitation).

We also do a parallel sort of each block of 10k rows returned so as to encourage sequential reads against the heap.  Unless the user specified their own sort, then we stick with that.

Finall, if scores aren't needed, we ask ES to elide returning the "_score" field just to avoid some network/json decoding overhead.