fix issue #622:  recursive updates no longer cause errors
There's two primary changes here.

The first is that during an UPDATE, we postpone queuing the  ES BulkRequestCommand until we see the corresponding insert command via the index AM API.

The second change is that we now track a set of "in flight" ctids (these are ctids that are either on their way to ES or being waited on for ES to finish indexing), and if we see an update command for one of these, we defer it (and its corresponding index command) and play those in after all the other documents have successfully finished.  This essentially happens recursively until all deferred commands have been processed.

The latter adds a small bit of per-tuple overhead in all cases (including CREATE INDEX), but it seems worth the price to ensure things actually work!