Rewrite build system (#628)
This rewrites the ./docker-build-system/build.sh script to a Rust binary.

Combined with some other changes to the number of CPUs allocated to each docker run command, it now builds everything on my monster computer about 2x faster (right under 18m).

It also now understands the number of CPUs in your computer and doesn't default to 64. Additionally, it only builds 1 image per Dockerfile, unless that Dockerfile as a ARG PGVER line, and then it'll automatically build one per Postgres version (this is only true for the "Alpine" image so far.

It also properly handles a `^C` and does a proper `docker kill` on all the running containers.

Generally, this is to resolve Issue #611.