Postgres is not a plural possessive
According to the Chicago Manual of Style, only a plural possessive
should be spelled in the form s', e.g. the thieves' den.
So Postgres' is not correct and Postgres's is preferred.

The AP stylebook disagrees for proper nouns, but
this README uses the Oxford comma, which AP also recommends against.