Working on issue #766:  improve hit highlighting performance
This does quite a bit of work to retool ZDB's hit highlighting functions (specifically `zdb.highlight_document()`) to be much faster.

A degenerate test case I came up with has gone from 5 minutes(!) to about 1.28 seconds on my laptop.

The major things that have changed are:

- Provide rust implementations of the "standard" and "fulltext_with_shingles" Elasticsearch analyzers so we can avoid setting text to Elasticsearch for tokenization
- Make these as zero-copy as possible
- Optimize the common case for "phrase queries" to be O(N * log N) instead of O(N^2)
- Optimize the various "find_zdb_index" paths (this was needlessly slow due to copying the large ZDBIndexOptionsInternal structure)

Also of note, the `zdb.highlight_document()` and `zdb.highlight_proximity()` functions have a new final argument, whose default is `true` to indicate if the highlighting results should be de-duplicated.  Setting it `false` can provide a bit more information about the number of times a specific token was matched if there's an overlapping proximity match within the same query clause.

There's still O(N^2) behavior for proximity queries with a distance greater than zero (in order or not), which will be addressed in subsequent commits.

All tests are passing...