fix issue #770
This adds a few more allowed time formats for the `time with(out) time zone` types.

It also adds a new function named `zdb.reapply_mapping(regclass)` that can be used to reapply the index mapping when a "mapping thing" changes, such as a row in the `zdb.type_mappings` table.