during indexing, move serialization for Date, Time, and TimeWithTimeZone to the main thread.
pgx 0.5.x changed the stringification of those types to use Postgres, and we cannot call Postgres from our background builder threads.

There'll likely be a little indexing performance hit for tables with lots of columns of these types (especially with arrays of these types), but that's better than crashing postgres.