Upgrade to pgrx 0.9.2.
Technically, this upgrades us to pgrx:develop because the process of upgrading here found a few bugs in pgrx 0.9.2 which are not yet released.

There was a bit of attention to date/time handling here, specifically converting the values to json.  A few of the existing regression tests have been updated to account for the changes.