Serialize VACUUM requests (#849)
Previously, the (auto)VACUUM process would send everything it needs to
do to Elasticsearch in a single `_bulk` request.  This could lead to
those individual actions happening out-of-order, causing the VACUUM
process to raise an ERROR and then leaving the index in an inconsistent
state.

This serializes the requests into 4 individual `_bulk` requests.