Fix issue #863
This adjusts hit highlighting to only remove "shingle" hits during final hit de-duplication if the "shingle" hit is not unique relative to the current hit being emitted.

This does adjust the expected output for issue #812, but after discussion with @Shigoki, we believe this is the correct behavior.

Also updates pgrx to 0.9.8.