Fix issue #867
The unary `NOT` operator is now pushed down into `Expr::Linked` nodes which is where it should be.  This ensures we're not retrieving values from ES that we shouldn't, which may not be properly filtered if the NOT was on the outside.

This does change the test around issue #679, but that was fixed incorrectly in the first place, which lead to this bug.